homework-12/src/main/java/ru/otus/spring/homework12/services/AuthorServiceImpl.java
package ru.otus.spring.homework12.services;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.otus.spring.homework12.repositories.AuthorRepository;
import ru.otus.spring.homework12.models.Author;

import java.util.List;

@RequiredArgsConstructor
@Service
public class AuthorServiceImpl implements AuthorService {
    private final AuthorRepository authorRepository;

    @Override
    public List<Author> findAll() {
        return authorRepository.findAll();
    }
}
